services/routes/requirements.txt
Flask
Flask-Caching
redis
pandas
gunicorn
psycopg2-binary
pytest
pytest-cov
pytest-mock
coverage
pytest-flask